requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
newsapi-python==0.2.6
requests==2.24.0
SQLAlchemy==1.3.18
urllib3==1.25.9
Werkzeug==1.0.1
